pyproject.toml
[tool.poetry]
name = "admin-service"
version = "0.1.0"
description = ""
authors = ["Grzegorz Pustułka <grzegorzpustulka@onet.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
uvicorn = "^0.27.0.post1"
fastapi = "^0.109.0"
email-validator = "^2.1.0.post1"
bcrypt = "^4.1.2"
secure-smtplib = "^0.1.1"
python-multipart = "^0.0.7"
psycopg2-binary = "^2.9.9"
hypercorn = "^0.16.0"
pika = "^1.3.2"
python-jwt = "^4.1.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
pre-commit = "^3.6.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
